I removed screenshot function and add another.
